Add files via upload
Kidofonics is a web Application intended to teach a visually impaired child to learn phonics to better children to learn reading better. The keyboard has braille alphabet letters and when a key(braille letter is pressed, a sound is heard by the child and it helps the child to learn to read better.